requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
tweepy>=4.14.0
asyncio>=3.4.3
streamlit>=1.28.0
pytest>=7.4.0
pytest-asyncio>=0.21.1
python-json-logger>=2.0.7 